Improve installer app structure
