+++ apps/com_pinoox_installer/router/actions.php
@@ -15,5 +15,7 @@ use function Pinoox\Router\{action};
 use App\com_pinoox_installer\Controller\MainController;
 use Pinoox\Component\Helpers\HelperHeader;
 
-action('home', [MainController::class,'home']);
-action('pinooxjs', [MainController::class,'pinooxjs']);
+action('home', [MainController::class, 'home']);
+action('pinooxjs', function () {
+    return View::response('pinoox', [], 'application/javascript', 'UTF-8');
+});
